move `winFindEditorPath` to new module
